Added get_by_user_id

+++ capstonebackend/reminders/views.py
@@ -10,46 +10,45 @@ from .serializers import ReminderSerializer
 from django.contrib.auth.models import User
 
 
-# Create your views here.
-class ReminderList(APIView):
+# Create your views here
 
-    permission_classes = [IsAuthenticated]
+@api_view(['GET'])
+@permission_classes([AllowAny])
 
-    def get(self, request):
-        reminders = Reminder.objects.all()
-        serializer = ReminderSerializer(reminders, many=True)
+# Gets all reminders from every user, probably wont be utalized
+def get_all_reminders(request):
+    reminders = Reminder.objects.all()
+    serializer = ReminderSerializer(reminders, many=True)
+    return Response(serializer.data)
+
+
+    
+@api_view(['GET','POST','PUT','DELETE'])
+@permission_classes([IsAuthenticated])
+
+#Only available if the user is logged in
+def user_reminders(request):
+    if request.method == 'GET':
+        reminder = Reminder.objects.filter(user_id=request.user.id)
+        serializer = ReminderSerializer(reminder, many=True)
         return Response(serializer.data)
 
-    def post(self, request):
+    if request.method == 'POST':
         serializer = ReminderSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(user=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-class ReminderDetail(APIView):
-
-    def get_object(self, pk):
-        try:
-            return Reminder.objects.get(pk=pk)
-        except Reminder.DoesNotExist:
-            raise Http404
-
-    def get(self, request, pk):
-        reminder = self.get_object(pk)
-        serializer = ReminderSerializer(reminder)
-        return Response(serializer.data)
-
-    def put(self, request, pk):
-        reminder = self.get_object(pk)
+    
+    if request.method == 'PUT':
+        reminder = Reminder.objects.filter(user_id=request.user.id)
         serializer = ReminderSerializer(reminder, data=request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(user=request.user)
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, pk):
-        reminder = self.get_object(pk)
+    if request.method == 'DELETE':
+        reminder = Reminder.objects.filter(user_id=request.user.id)
         reminder.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
